simplify login modules
