+++ CHANGELOG.md
@@ -9,9 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- add support for logging in with ENTER key
+- add support for logging-in with ENTER key
 
 ### Changes
+
+- set is_loading to True on_load for search and ingest states
 
 ### Deprecated
