+++ pyproject.toml
@@ -10,7 +10,7 @@ requires-python = ">=3.11,<3.12"
 dependencies = [
     "async-lru>=2",
     "fastapi>=0.115",
-    "mex-common>=0.62",
+    "mex-common>=0.62.1,<0.63",
     "pydantic>=2",
     "pyyaml>=6",
     "reflex>=0.6,<0.7",
@@ -30,7 +30,7 @@ optional-dependencies.dev = [
     "pytest>=8",
     "ruff>=0.11",
     "sphinx>=8",
-    "types-ldap3>=2.9.13.20250516",
+    "types-ldap3>=2",
     "types-PyYAML>=6",
     "types-requests>=2",
 ]
